+++ Migrations/UserContextModelSnapshot.cs
@@ -44,16 +44,16 @@ namespace calculadora_api.Migrations
                     b.Property<string>("indiceDB")
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<float>("indiceDataBase")
+                    b.Property<double>("indiceDataBase")
                         .HasColumnType("real");
 
-                    b.Property<float>("indiceDataBaseAtual")
+                    b.Property<double>("indiceDataBaseAtual")
                         .HasColumnType("real");
 
                     b.Property<string>("infoParaCalculo")
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<float>("lancamentos")
+                    b.Property<double>("lancamentos")
                         .HasColumnType("real");
 
                     b.Property<string>("tipoLancamento")
@@ -62,10 +62,10 @@ namespace calculadora_api.Migrations
                     b.Property<string>("ultimaAtualizacao")
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<float>("valorDevedor")
+                    b.Property<double>("valorDevedor")
                         .HasColumnType("real");
 
-                    b.Property<float>("valorDevedorAtualizado")
+                    b.Property<double>("valorDevedorAtualizado")
                         .HasColumnType("real");
 
                     b.HasKey("Id");
@@ -87,7 +87,7 @@ namespace calculadora_api.Migrations
                         .HasColumnType("nvarchar(max)");
 
                     b.Property<double>("valor")
-                        .HasColumnType("float");
+                        .HasColumnType("double");
 
                     b.HasKey("Id");
 
@@ -143,7 +143,7 @@ namespace calculadora_api.Migrations
                         .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
-                    b.Property<float>("amortizacao")
+                    b.Property<double>("amortizacao")
                         .HasColumnType("real");
 
                     b.Property<string>("contractRef")
@@ -164,10 +164,10 @@ namespace calculadora_api.Migrations
                     b.Property<string>("indiceDV")
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<float>("indiceDataCalcAmor")
+                    b.Property<double>("indiceDataCalcAmor")
                         .HasColumnType("real");
 
-                    b.Property<float>("indiceDataVencimento")
+                    b.Property<double>("indiceDataVencimento")
                         .HasColumnType("real");
 
                     b.Property<string>("infoParaAmortizacao")
@@ -179,28 +179,28 @@ namespace calculadora_api.Migrations
                     b.Property<string>("nparcelas")
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<float>("parcelaInicial")
+                    b.Property<double>("parcelaInicial")
                         .HasColumnType("real");
 
                     b.Property<string>("status")
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<float>("subtotal")
+                    b.Property<double>("subtotal")
                         .HasColumnType("real");
 
                     b.Property<string>("tipoParcela")
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<float>("totalDevedor")
+                    b.Property<double>("totalDevedor")
                         .HasColumnType("real");
 
                     b.Property<string>("ultimaAtualizacao")
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<float>("valorNoVencimento")
+                    b.Property<double>("valorNoVencimento")
                         .HasColumnType("real");
 
-                    b.Property<float>("valorPMTVincenda")
+                    b.Property<double>("valorPMTVincenda")
                         .HasColumnType("real");
 
                     b.HasKey("Id");
